Remove make methods from builder interface.
This removes a hard dependency between instance type and builder. Without this dependency, an instance type can be built from different builders. Note that we could change the implementation of the make methods, however, they have mostly been replaced with direct builder instancing, so their value is questionable. If they are required, we would simply not be able to automatically deduct the builder type from the instance type and instead would have to do it the other way around (i.e. deduct the instance type from the builder type).