Disable (potential) false positive barrier error.
This only happens when ending a render pass that draws directly to a swap-chain back buffer. I reported this to MSFT using discord and they are investigating. Since all barriers of the app are now new enhanced barriers, we should be safe to ignore this error in the long run.