Core: Add alias for handle type in a resource
This adds a handle_type of type THandle in every class inheriting Resource.
It allows to get the type of the underlying backend resource when having the
backend type as template.
It avoids a runtime check by backend name with ifdefs for platforms.

This will be used in the next commits to create the windows surface on linux.